+++ dataseries/categorical.lua
@@ -3,7 +3,6 @@ local Dataseries = params[1]
 
 local argcheck = require "argcheck"
 local doc = require "argcheck.doc"
-local tds = require "tds"
 
 doc[[