+++ NEWS.md
@@ -3,6 +3,11 @@ News file for torch-dataframe
 
 Version: development
 --------------------
+* The get_max_value/get_min_value use torch.max/min when no missing data is present in the column
+* Fixed upgrade_frame bug
+* Fixed bug with saving CSV-files when they contain boolean values
+* Added faster torch.Tensor functions to fill/stat functions for speed
+* Added mutate function to Dataseries
 
 Version: 1.6
 --------------------